feat: rename ResultCopyButton to CopyButton and use in DomEvents
