+++ src/components/icons/SuccessIcon.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+function SuccessIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="currentColor"
+      width="16"
+      height="16"
+      viewBox="0 0 512 512"
+    >
+      <path d="M256 8a248 248 0 100 496 248 248 0 000-496zm0 48a200 200 0 110 400 200 200 0 010-400m140 130l-22-22c-5-5-13-5-17-1L215 304l-59-61c-5-4-13-4-17 0l-23 23c-5 5-5 12 0 17l91 91c4 5 12 5 17 0l172-171c5-4 5-12 0-17z" />
+    </svg>
+  );
+}
+
+export default SuccessIcon;
